Command Line Arguments handling
Adding a real command line arguments parsing and handling mechanisms,
thanks to QCommandLine [http://xf.iksaif.net/dev/qcommandline.html].
Adding a setDefaultParameter() method for the Parameters object, setting
defaults values.
Adding a duration command line argument to allow user to set the length
(in minutes) of the presentation. Default is 20 minutes.
Adding an emergency command line argument to allow user to set the date
(in minutes) when it reaches the critical state that the speaker has not
long time before it ends. Default is 2 minutes.
Adding a slide-width command line argument to allow user to change the
width of the main slide displayed on mainscreen. Default is 70%.