Fixing Beamer speaker notes handling
Previously, only the current speaker's note was stored in a QString of
PDFModel. This value was set by renderPdfPage(), which is now not always
called (this is normal). This resulted in erratic appearance of the
notes. We now store notes (which are expected to be pretty short texts)
as a QHash<int, QString>. Thus, we can lazy-fill this hashmap when
renderPdfPage() is called, and retrieve correctly the value associated
with the current page.