Introducing a working Options Window
To ease changing options of the presentation, and mainly for those who
dislikes command line interface, introduction of a window to set some of
those parameters. Not all parameters are handled, but you can change the
following:
 - main screen current slide width
 - presentation duration
 - presentation emergency notification
 - main screen identifier
 - projector screen identifier

Screen identifier switching needed several modifications in the way
screens are handled. It's now better, safer and smoother :).

Key binding for this new tool is 'o'. Of course, french locale is
updated to those new changes.