Better live windows switching
Now reached a cleaner state, where both windows are at the correct size,
and where projector's rendering is scaled correctly, and image is
centered. Also needed to change the locale date format from Long to
Short, due to (pretty) large size it has: this was the reason why main
screen informations weren't displayed correctly when resized on a small
screen.