[FIX] Handling of first/last slide when going previous/next
There was a bug when playing at borders of the presentation:
 - going at last slide, then pressing Next, then going back
or
 - going at first slide, then pressing Previous, then going forward

In both case, the current page displayed on the projector screen wasn't
the expected one. This is due to the handling of next/previous page goto
code. Fix is simply check if we're at some boundaries (either first
slide or last slide) and if it's the case, just do nothing:
 - going previous when at first slide is insane
 - going next when at last slide is insane too.