Adding Left/Right mouse button handling
Being able to switch slides from keyboard is a nice feature, but
sometimes you might want to be able to control them from your mouse.
This commit enables the user to forward/backward slides with the
left/right mouse button. Warning: it's only working is you're on the
main screen. When navigating mouse to the secondary screen, it's not
enabled. Is it really needed ?