Splitting screen saver inhibiters
Reverting static methods and exception, using ScreenSaverInhibiters
class as an interface which has to be implemented by all inhibiters.
Each one will be tested and used if available. Testing is done by
implementing the canHandle() methods, and requires checking if the
method can be applied here (e.g. Freedesktop inhibiter will test if
there is the needed DBus interface available).