Preliminary Video Playback with LibVLC
This is only proof of concept code, but it plays an attached file by
storing it in a temporary file, then using LibVLC to play it on the
widget. Still need to refactor so that playback is done outside of
any widget (pdfmodel for example) and only output frames is sent to
widgets. Then, we will need to avoid storing in a temprorary file.