fix: Media file detection from Poppler
Poppler 0.16 was bugged and put the mime type/contents in the
description field. This has been fixed in 0.18. But the movie15 does not
correctly set the SubType value of the EmbeddedFile stream, hence the
isMediaFile() test is broken with poppler 0.18 and movie15-generated
files:
 - description() is empty as it should be
 - mimeType() is empty but it should not be
Work around is to look a the contents() method of the
FileAttachmentAnnotation, which contains the same string.